Refactor Stats models
