+++ src/server/services/lifetimeStats/index.ts
@@ -1,10 +1,8 @@
 import { SERVICE_Call__getAllReposWithCommitCounts } from '@/server/services/reposAndCommitCounts'
-import { SHARED_APIFields__LifetimeStats, SHARED_Model__LifetimeStats } from '@/shared/models'
+import { SHARED_APIFields__LifetimeStats } from '@/shared/models/apiFields'
 import { SHARED_Model__AllLanguageStats } from '@/shared/models/models/Language'
-import {
-    SHARED_Model__RepoCommitCountStats,
-    SHARED_Model__RepoWithCommitCountsAndLanguages,
-} from '@/shared/models/models/Repos'
+import { SHARED_Model__RepoWithCommitCountsAndLanguages } from '@/shared/models/models/Repos'
+import { SHARED_Model__LifetimeStats, SHARED_Model__RepoCommitCountStats } from '@/shared/models/models/Stats'
 
 const sharedRepoReduceLanguageStats = (
     repos: SHARED_Model__RepoWithCommitCountsAndLanguages[],
