+++ src/shared/models/apiFields/index.ts
@@ -1,3 +1,4 @@
+import { SHARED_Model__LifetimeStats } from '../models/Stats'
 import { SHARED_Model__UserCardData } from '../models/UserCard'
 
 export interface SHARED_APIFields__BASE {
@@ -13,4 +14,8 @@ export type SHARED_APIFields__ValidateToken = SHARED_APIFields__BASE
 
 export interface SHARED_APIFields__GetUserCard extends SHARED_APIFields__BASE {
     userCard?: SHARED_Model__UserCardData
+}
+
+export interface SHARED_APIFields__LifetimeStats extends SHARED_APIFields__BASE {
+    lifetimeStats?: SHARED_Model__LifetimeStats
 }
